Cleanup: make thermostat representation clearer.
Add a thermostat module with a Thermostat class to make the
implementation clearer.  Get rid of state machine code in favour of
simpler if/then clauses since the implementation can be short enough for
this to be reasonable.

Thermostat objects are akin to a physical thermostat: they take a
temperature input and decide how to act upon it, and have a connection
to a boiler that they can turn on or off.