Model temperature sensors separately and support reading history.
Previously, zones recorded the temperature they saw and there was a
direct mapping from zone to the sensor MQTT topic.  This change splits
them apart so there can be more sensors than zones, and the sensor is
abstracted from the zone.  This will allow different sensor types in
future, and with further separation could move the sensor to be used for
tracking into the schedule, allowing different sensors to be used for
the same zone at different times of day.

Immediately, this also includes changes to track sensor history.  A new
set of API endpoints, /sensor*, can be used to post and retrieve sensor
readings.  In the backend, the history of readings that are posted is
stored.

NOTE: This includes critical updates to scheduler.sql that should have
been included in previous commits.