Add HTML Basic Auth to the scheduler app.
Previously this would be provided by the webserver, but by adding it
directly into the app we can also support other types of authentication.
Note that this implies a different username, currently a device ID,
should be used for devices communicating with the service.

This DOES NOT include granular permissions or any kind of multi-tenancy.
any authenticated endpoint gets full access to the API.