Allow login via access token if valid client secret presented.
This allows us to support Alexa integration (in a
test-only/non-releaseable way) without having to host an OAuth server
ourselves.  We continue to use the Google SOO method via Alexa, so will
receive Google access tokens to authenticate with.  Requiring the client
secret for this kind of authentication limits it to cases where we trust
the client, to protect against access tokens from authentication with
other applications being used.