Support adding a prefix to all URLs.
Previously I'd rewritten the URL coming to the application, however since it
generates (via the flask-restx framework) some absolute URL this was
insufficient.  Instead, make it so I can set the base URL as a config parameter
and pass this through.  Fix up documentation by re-registering the endpoint to
the blueprint.